+++ src/ebsclient.h
@@ -5,8 +5,8 @@
 
 #define FACE_TEMPLATE_SIZE	512
 
-int get_face_template(const unsigned char* image_data, const unsigned int image_data_len,
-	void* template_buf, const unsigned int template_buf_size);
+int get_face_template(const unsigned char* image_data, const size_t image_data_len,
+	void* template_buf, const size_t template_buf_size);
 
 float cmp_face_template(void* tmp1, void* tmp2);
